Code cleanup, use settings in Diag Feature
