+++ src/DiagnosticsServiceHarness/project.json
@@ -7,7 +7,6 @@
     "DiagnosticsHarnessMessages": { "target": "project" },
     "StructureMap": "4.4.3",
     "StructureMap.Microsoft.DependencyInjection": "1.3.0",
-    "PeterKottas.DotNetCore.WindowsService": "1.0.16",
     "Microsoft.NETCore.App": {
       "version": "1.1.0",
       "type": "platform"
