+++ src/Jasper.Diagnostics/DiagnosticsMiddlewareExtensions.cs
@@ -1,12 +1,9 @@
 using System;
-using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Primitives;
-using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.FileProviders.Embedded;
 using JasperBus;
 using Jasper.Remotes.Messaging;
@@ -44,8 +41,6 @@ namespace Jasper.Diagnostics
 
             var hub = app.ApplicationServices.GetService<IMessagingHub>();
             var manager = app.ApplicationServices.GetService<ISocketConnectionManager>();
-
-            Console.WriteLine("manager? {0},{1}", manager != null, hub != null);
 
             app.MapWebSocket($"{options.BasePath}/ws",
                 new SocketConnection((socket, text) => {
