+++ src/Jasper.Diagnostics/DiagnosticsServer.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using Jasper.Diagnostics.StructureMap;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -9,44 +10,7 @@ using StructureMap;
 
 namespace Jasper.Diagnostics
 {
-    public static class ServiceCollectionExtensions
-    {
-        public static IWebHostBuilder UseStructureMap(this IWebHostBuilder builder, IContainer container)
-        {
-            return builder.ConfigureServices(services => services.AddStructureMap(container));
-        }
-
-        public static IServiceCollection AddStructureMap(this IServiceCollection services, IContainer container)
-        {
-            return services.AddSingleton<IServiceProviderFactory<IContainer>>(new StructureMapServiceProviderFactory(container));
-        }
-    }
-
-    public class StructureMapServiceProviderFactory : IServiceProviderFactory<IContainer>
-    {
-        public StructureMapServiceProviderFactory(IContainer container)
-        {
-            Container = container;
-        }
-
-        private IContainer Container { get; }
-
-        public IContainer CreateBuilder(IServiceCollection services)
-        {
-            var registry = Container ?? new Container();
-
-            registry.Populate(services);
-
-            return registry;
-        }
-
-        public IServiceProvider CreateServiceProvider(IContainer container)
-        {
-            return new StructureMapServiceProvider(container);
-        }
-    }
-
-    internal class Startup
+    internal class DiagnosticsServerStartup
     {
         public void ConfigureServices(IServiceCollection services)
         {
@@ -58,12 +22,8 @@ namespace Jasper.Diagnostics
 
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
-            app.UseDiagnostics(_ =>
-            {
-                // _.Mode = DiagnosticsMode.Development;
-                // _.AuthorizeWith = context => context.User.HasClaim("admin", "true");
-                _.WebsocketPort = 5250;
-            });
+            var settings = app.ApplicationServices.GetService<DiagnosticsSettings>();
+            app.UseDiagnostics(settings);
 
             app.Run(async http =>
             {
@@ -77,35 +37,20 @@ namespace Jasper.Diagnostics
     public class DiagnosticsServer : IDisposable
     {
         private IWebHost _host;
+        private DiagnosticsSettings _settings;
 
-        public void Start(int port, IContainer container)
+        public void Start(DiagnosticsSettings settings, IContainer container)
         {
-            var hrm = container.WhatDoIHave();
-            Console.WriteLine(hrm);
+            _settings = settings;
 
+            var url = $"http://localhost:{settings.WebsocketPort}";
+            Console.WriteLine($"Diagnostics listening on {url}");
             _host = new WebHostBuilder()
                 .UseKestrel()
                 .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseUrls($"http://localhost:{port}")
+                .UseUrls(url)
                 .UseStructureMap(container)
-                .UseStartup<Startup>()
-//                .Configure(app =>
-//                {
-//                    app.UseDiagnostics(_ =>
-//                    {
-//                        // _.Mode = DiagnosticsMode.Development;
-//                        // _.AuthorizeWith = context => context.User.HasClaim("admin", "true");
-//                        _.WebsocketPort = port;
-//                    });
-
-//                    app.Run(async http =>
-//                    {
-//                        http.Response.StatusCode = 200;
-//                        http.Response.ContentType = "text/plain";
-//                        // await http.Response.WriteAsync($"Nothing to see here at {DateTime.Now}.");
-//                        await http.Response.WriteAsync(hrm);
-//                    });
-//                })
+                .UseStartup<DiagnosticsServerStartup>()
                 .Build();
 
             _host.Start();
