+++ src/Jasper.Diagnostics/Messages/DiagnosticsListener.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using JasperBus.Model;
 using Jasper.Remotes.Messaging;
@@ -21,58 +19,8 @@ namespace Jasper.Diagnostics.Messages
 
         public void Receive(RequestInitialData message)
         {
-            var chains = _graph.Chains.OrderBy(c => c.TypeName).Select(c => ChainModel.For(c));
+            var chains = _graph.Chains.OrderBy(c => c.TypeName).Select(ChainModel.For);
             _client.Send(new InitialData(chains));
         }
-    }
-
-    public class RequestInitialData : ClientMessage
-    {
-        public RequestInitialData()
-            : base("request-initial-data")
-        {
-        }
-    }
-
-    public class InitialData : ClientMessage
-    {
-        public InitialData(IEnumerable<ChainModel> chains)
-            : base("initial-data")
-        {
-            Chains = chains.ToArray();
-        }
-
-        public ChainModel[] Chains { get; }
-    }
-
-    public class ChainModel
-    {
-        public MessageTypeModel MessageType { get; set; }
-        public string GeneratedTypeName { get; set; }
-        public string Description { get; set; }
-        public string SourceCode { get; set; }
-
-        public static ChainModel For(HandlerChain chain)
-        {
-            return new ChainModel
-            {
-                MessageType = new MessageTypeModel(chain.MessageType),
-                GeneratedTypeName = chain.TypeName,
-                Description = chain.ToString(),
-                SourceCode = chain.SourceCode
-            };
-        }
-    }
-
-    public class MessageTypeModel
-    {
-        public MessageTypeModel(Type type)
-        {
-            Name = type?.Name;
-            FullName = type?.FullName;
-        }
-
-        public string Name { get; }
-        public string FullName { get; }
     }
 }
