+++ src/Jasper.Diagnostics/Messages/MessageTypeModel.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace Jasper.Diagnostics.Messages
+{
+    public class MessageTypeModel
+    {
+        public MessageTypeModel(Type type)
+        {
+            Name = type?.Name;
+            FullName = type?.FullName;
+        }
+
+        public string Name { get; }
+        public string FullName { get; }
+    }
+}
