+++ src/Jasper.Diagnostics/Messages/RequestInitialData.cs
@@ -1,0 +1,12 @@
+using Jasper.Remotes.Messaging;
+
+namespace Jasper.Diagnostics.Messages
+{
+    public class RequestInitialData : ClientMessage
+    {
+        public RequestInitialData()
+            : base("request-initial-data")
+        {
+        }
+    }
+}
