+++ src/Jasper/project.json
@@ -25,11 +25,11 @@
     "Microsoft.CodeAnalysis.CSharp": "2.0.0-rc4",
     "Microsoft.CodeAnalysis.Scripting": "2.0.0-rc4",
     "Newtonsoft.Json": "9.0.1",
-    "Microsoft.Extensions.Caching.Memory": "1.1.1",
-    "Microsoft.Extensions.Configuration": "1.1.1",
-    "Microsoft.Extensions.Configuration.Abstractions": "1.1.1",
-    "Microsoft.Extensions.Configuration.Binder": "1.1.1",
-    "Microsoft.Extensions.Configuration.Json": "1.1.1",
+    "Microsoft.Extensions.Caching.Memory": "1.1.0",
+    "Microsoft.Extensions.Configuration": "1.1.0",
+    "Microsoft.Extensions.Configuration.Abstractions": "1.1.0",
+    "Microsoft.Extensions.Configuration.Binder": "1.1.0",
+    "Microsoft.Extensions.Configuration.Json": "1.1.0",
     "StructureMap": "4.4.3"
   },
   "frameworks": {
