+++ src/JasperBus.Tests/project.json
@@ -4,7 +4,7 @@
     "debugType": "portable"
   },
   "dependencies": {
-    "System.Runtime.Serialization.Primitives": "4.1.1",
+    "System.Runtime.Serialization.Primitives": "4.3.0",
     "xunit": "2.2.0-beta2-build3300",
     "dotnet-test-xunit": "2.2.0-preview2-build1029",
     "JasperBus": { "target": "project" },
